requirements.txt
numpy
gymnasium
PyYAML
agents
scikit-learn
matplotlib

